More infrastructure work, include MAX_WORKER_ATTEMPTS and ShouldDeleteFromReceiving

+++ ingest/cleanup.go
@@ -38,11 +38,13 @@ func (c *Cleanup) Run() (fileCount int, errors []*service.ProcessingError) {
 		}
 	}
 	if len(errors) == 0 {
-		// TODO: IngestObject should probably have a flag describing
-		// whether the original bag should be deleted from storage.
-		err := c.deleteFromReceiving()
-		if err != nil {
-			errors = append(errors, c.Error(c.IngestObject.Identifier(), err, false))
+		if c.IngestObject.ShouldDeleteFromReceiving {
+			err := c.deleteFromReceiving()
+			if err != nil {
+				errors = append(errors, c.Error(c.IngestObject.Identifier(), err, false))
+			}
+		} else {
+			c.Context.Logger.Warning("WorkItem %d: Not deleting %s/%s from receiving bucket because ShouldDeleteFromReceiving = false", c.WorkItemID, c.IngestObject.S3Bucket, c.IngestObject.S3Key)
 		}
 	}
 	return fileCount, errors
